tests/requirements.txt
pandas
requests
awswrangler
flatten_json
charset-normalizer
pytest-asyncio